放弃使用spring cache 集成redis . 更改为 redisTemplate ,并对其封装了接口。

+++ yph-system/src/main/java/com/yph/controller/sys/SysMenuController.java
@@ -41,7 +41,7 @@ public class SysMenuController {
     @ResponseBody
     @RequestMapping("saveSysMenu")
     public CommonResult saveSysMenu(SysMenu sysMenu) {
-        int result = sysMenuService.saveSysMenu(sysMenu);
+        Long result = sysMenuService.saveSysMenu(sysMenu);
         if (result < 1) {
             return CommonResult.ERROR("保存菜单失败!");
         }
@@ -155,4 +155,6 @@ public class SysMenuController {
         List<ZtreeVo> list = sysMenuService.findListByZtree(params);
         return CommonResult.SUCCESS(list);
     }
+
+
 }
